Update main.py
implemented options for CLI args using the new 3.10+ match case (to mimic switch-case). Also used a splitting method so I can avoid having to run multiple if-checks for each additional parameter added. Implemented the --verbose option for use with the update command to show old and new prices as opposed to the simple progress bar.